+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Add news component FormField
 - Add store pages methods to add and remove
 - Add ButtonUi component
+- Add CollapseUi component
 ### Changed
 - Changed the layoutAdmin grid to prevent jump of content on transition
 ### Deprecated